Refactoring data access class
Вынесены в астрактный класс функционал транзакций и соединение с БД т.к. это поле и транзакции присутствуют в любом классе работы с БД.
